- Added OperationNormalizer and minor refactor on JsonDiff
