+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+#Changelog
+
+## 2018-04-14 [0.4.5]
+
+- [Added] Changelog.
+- [Added] ComparisonCriteria, now you can set custom equality rules to JsonDiff.
+- [Changed] Now JsonDiff is instantiable, you can create a new JsonDiff instance using the `JsonDiff.builder()`, see README.md for more details.
